Remove ASR work, it does not belong in this PR anymore.
